custom_components/ev_load_balancer/__init__.py
"""EV Load Balancer — dynamisk lastbalansering för Home Assistant."""

import logging

from homeassistant.config_entries import ConfigEntry
from homeassistant.core import HomeAssistant

from .const import DOMAIN

_LOGGER = logging.getLogger(__name__)


async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
    """Sätt upp integrationen från en config entry."""
    hass.data.setdefault(DOMAIN, {})
    hass.data[DOMAIN][entry.entry_id] = {}
    return True


async def async_unload_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
    """Avinstallera integrationen och rensa hass.data."""
    hass.data[DOMAIN].pop(entry.entry_id, None)
    return True
